+++ src/main/java/sic/modelo/CuentaCorrienteProveedor.java
@@ -3,6 +3,7 @@ package sic.modelo;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import javax.persistence.Entity;
@@ -30,11 +31,11 @@ public class CuentaCorrienteProveedor extends CuentaCorriente implements Seriali
         return proveedor.getRazonSocial();
     }
 
-    public CuentaCorrienteProveedor() {
-    }
+    public CuentaCorrienteProveedor() {}
 
     public CuentaCorrienteProveedor(long idCuentaCorriente, boolean eliminada, Date fechaApertura, Empresa empresa,
-            double saldo, Date fechaUltimoMovimiento, List<RenglonCuentaCorriente> renglones, Proveedor proveedor) {
+            BigDecimal saldo, Date fechaUltimoMovimiento, List<RenglonCuentaCorriente> renglones, Proveedor proveedor) {
+        
         super(idCuentaCorriente, eliminada, fechaApertura, empresa, saldo, fechaUltimoMovimiento, renglones);
         this.proveedor = proveedor;
     }
